+++ GetOffMyLawn/Patches/PrivateAreaPatch.cs
@@ -8,8 +8,12 @@ using static PluginConfig;
 static class PrivateAreaPatch {
   [HarmonyPostfix]
   [HarmonyPatch(nameof(PrivateArea.Interact))]
-  static void InteractPostfix(ref PrivateArea __instance) {
-    if (IsModEnabled.Value && __instance && __instance.IsEnabled() && __instance.m_piece.IsCreator()) {
+  static void InteractPostfix(PrivateArea __instance) {
+    if (IsModEnabled.Value
+        && RepairPiecesOnWardActivation.Value
+        && __instance
+        && __instance.IsEnabled()
+        && __instance.m_piece.IsCreator()) {
       PieceUtils.RepairPiecesInRadius(__instance.transform.position, __instance.m_radius);
     }
   }