+++ GetOffMyLawn/Patches/WearNTearPatch.cs
@@ -9,32 +9,32 @@ static class WearNTearPatch {
   public static readonly long PieceHealthDamageThreshold = 100_000L;
 
   [HarmonyPrefix]
-  [HarmonyPatch(nameof(WearNTear.ApplyDamage))]
-  static bool ApplyDamagePrefix(ref WearNTear __instance, ref bool __result, ref float damage) {
-    if (!IsModEnabled.Value || !EnablePieceHealthDamageThreshold.Value) {
-      return true;
-    }
-
-    float health = __instance.m_nview.m_zdo.GetFloat(ZDOVars.s_health, __instance.m_health);
-
-    if (health <= 0f) {
-      __result = false;
-      return false;
-    } else if (health >= PieceHealthDamageThreshold) {
-      __result = false;
+  [HarmonyPatch(nameof(WearNTear.ApplyDamage), new[] { typeof(float), typeof(HitData) })]
+  static bool ApplyDamagePrefix(WearNTear __instance, float damage, HitData hitData, ref bool __result) {
+    if (IsModEnabled.Value && EnablePieceHealthDamageThreshold.Value) {
+      __result = ApplyDamageDelegate(__instance, damage, hitData);
       return false;
     }
 
-    health -= damage;
-    __instance.m_nview.m_zdo.Set(ZDOVars.s_health, health);
+    return true;
+  }
+
+  public static bool ApplyDamageDelegate(WearNTear wearNTear, float damage, HitData hitData) {
+    float health = wearNTear.m_nview.m_zdo.GetFloat(ZDOVars.s_health, wearNTear.m_health);
+
+    if (health <= 0f || health >= PieceHealthDamageThreshold) {
+      return false;
+    }
+
+    wearNTear.m_nview.m_zdo.Set(ZDOVars.s_health, health);
 
     if (health <= 0f) {
-      __instance.Destroy();
+      wearNTear.Destroy(hitData);
     } else {
-      __instance.m_nview.InvokeRPC(ZNetView.Everybody, "WNTHealthChanged", health);
+      ZRoutedRpc.s_instance.InvokeRoutedRPC(
+          ZRoutedRpc.Everybody, "RPC_HealthChanged", wearNTear.m_nview.m_zdo.m_uid, health);
     }
 
-    __result = true;
-    return false;
+    return true;
   }
 }